Initial rendering ot TextPopupElement
Moved WPF specific code to seperate file
